packer: simplify basin manifest
There is no a real need for separating different types of
content in separate fields, other than content in general
and sets.

What we really need is a way of telling whether something
needs to be indexed (e.g, a real content) and when not to
(e.g. resources).

This gives us extra flexibility to combine things around,
e.g, having videos that are index and videos that are not.

https://phabricator.endlessm.com/T19028